PDF Report (#13)
Pdf report utility function to support benchmarking output results rendering.

ALso created a small script generate_report.py that provide 2 usage, 1 that walks you over benchmark directory to choose a specific dataset and model to generate a report on, and onther one that accept a model_dir directly.
This was necessary to simplify working and testing with preexisting benchmark data.

---------

Co-authored-by: bru08 <ppabrno@gmail.com>